Move table rows and empty row to separate components. Move form handling logic up a component.
